+++ src/components/playlist/Menu/PlaylistMenu.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 
 import { useNoxSetting } from '@stores/useApp';
 import usePlaylist from '@hooks/usePlaylistRN';
-import PlaylistSettingsButton from '../Setting/PlaylistSettingsButton';
+import PlaylistSettingsButton from './PlaylistSettingsButton';
 import { PLAYLIST_ENUMS } from '@enums/Playlist';
 import { CopiedPlaylistMenuItem } from '../../buttons/CopiedPlaylistButton';
